Move toggle to kit
